+++ lib/lower/Bridge.cpp
@@ -10,6 +10,7 @@
 #include "fir/Dialect/FIRDialect.h"
 #include "fir/Dialect/FIROps.h"
 #include "fir/Dialect/FIRType.h"
+#include "fir/InternalNames.h"
 #include "flang/lower/ASTBuilder.h"
 #include "flang/lower/ConvertExpr.h"
 #include "flang/lower/ConvertType.h"
@@ -18,7 +19,6 @@
 #include "flang/lower/Mangler.h"
 #include "flang/lower/OpBuilder.h"
 #include "flang/lower/Runtime.h"
-#include "flang/optimizer/InternalNames.h"
 #include "flang/parser/parse-tree.h"
 #include "flang/semantics/tools.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
