+++ lib/optimizer/CodeGen.cpp
@@ -11,8 +11,8 @@
 #include "fir/Dialect/FIRDialect.h"
 #include "fir/Dialect/FIROps.h"
 #include "fir/Dialect/FIRType.h"
-#include "flang/optimizer/InternalNames.h"
-#include "flang/optimizer/KindMapping.h"
+#include "fir/InternalNames.h"
+#include "fir/KindMapping.h"
 #include "mlir/Conversion/StandardToLLVM/ConvertStandardToLLVM.h"
 #include "mlir/Conversion/StandardToLLVM/ConvertStandardToLLVMPass.h"
 #include "mlir/Dialect/AffineOps/AffineOps.h"
